requirements.txt
requests>=2.31.0
urllib3>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
paramiko>=3.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
